Update AnalyzeMATCH.m
Optimization of the sampling rate has been included after it was found that the HexaMotion platform drives different motion traces with different speeds. The user has to enter an initial sampling rate and the optimizer tries to find a better fit withing +- 5% of that sampling rate by maximizing the correlation of the resampled measured motion with the ground truth motion trace.